refactor(cmd): namespace command internals
Group the command parser, completion, dispatch, and resolver helpers under the
forge.cmd module family so the public forge.cmd facade stays stable while the
root lua/forge directory stops exposing those internal implementation files.
