+++ spec/command_spec.lua
@@ -26,10 +26,10 @@ local loaded_modules = {
   'forge',
   'forge.config',
   'forge.cmd',
-  'forge.cmd_complete',
-  'forge.cmd_complete_source',
-  'forge.cmd_dispatch',
-  'forge.cmd_resolve',
+  'forge.cmd.complete',
+  'forge.cmd.complete_source',
+  'forge.cmd.dispatch',
+  'forge.cmd.resolve',
   'forge.detect',
   'forge.issue',
   'forge.logger',
